requirements.txt
django
segno
pillow
gunicorn
whitenoise
fpdf2